Add support for proxy models
This introduces an optional argument to `query_string_for_field` for
explicitly specifying the model the field is defined on. This allows
for passing a proxy model along with the bound field which will alter
the join path that is constructed.

Read about proxy models in Django: https://docs.djangoproject.com/en/dev/topics/db/models/#proxy-models


Signed-off-by: Byron Ruth <b@devel.io>