Rename core tests with test_ prefix so django recognizes them
Django 1.6 uses a new test runner which requires tests to be in files
that start with "test". Django 1.5 and earlier would find these in the
tests/ directory but that is no longer the case so we add the test_
prefix which has no negative effects in Django <= 1.5 but allows the
test runner to find them in Django >= 1.6.

Signed-off-by: Don Naegely <naegelyd@email.chop.edu>